+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -152,7 +152,7 @@ class DataTableRoot<
   }
 
   componentDidUpdate(prevProps: any) {
-    const { data, selectedRows, columns } = this.asProps;
+    const { data, selectedRows, columns, loading } = this.asProps;
     if (prevProps.columns !== columns) {
       const cols = this.calculateColumnsFromConfig();
       this.columns = cols[0];
@@ -166,11 +166,31 @@ class DataTableRoot<
       }
     }
     if (prevProps.selectedRows !== selectedRows && selectedRows !== undefined) {
-      if (prevProps.selectedRows.length < data.length && selectedRows.length === data.length) {
+      const selectedRowsSet = new Set<UniqKeyType>(selectedRows);
+
+      const allChecked: UniqKeyType[] = [];
+      const allUnchecked: UniqKeyType[] = [];
+
+      this.flatRows.forEach((row) => {
+        if (selectedRowsSet.has(row[UNIQ_ROW_KEY])) {
+          allChecked.push(row[UNIQ_ROW_KEY]);
+        } else {
+          allUnchecked.push(row[UNIQ_ROW_KEY]);
+        }
+      });
+
+      if (allChecked.length === data.length) {
         this.setSelectAllMessage(true);
-      } else if (prevProps.selectedRows.length > 0 && selectedRows.length === 0) {
+      } else if (allUnchecked.length === data.length) {
         this.setSelectAllMessage(false);
       }
+    }
+    if (prevProps.loading !== loading) {
+      setTimeout(() => {
+        if (document.activeElement === document.body || (this.tableContainerRef.current && hasParent(document.activeElement, this.tableContainerRef.current))) {
+          this.tableRef.current?.focus();
+        }
+      }, 0);
     }
   }